feat(deploy): add Fabric script for deploying web_static archive
Added a Fabric script to deploy an archive to web servers. The script includes functions for generating a .tgz archive from the web_static folder and distributing it to the specified servers.

The do_pack function creates a timestamped archive of the web_static folder contents, while the do_deploy function uploads the archive, extracts its contents on the servers, updates symbolic links, and cleans up obsolete files.